completed

+++ app/Http/Controllers/IdeaController.php
@@ -27,6 +27,8 @@ class IdeaController extends Controller
             'content' => 'required | min:3 | max:240',
         ]);
 
+        $validated['user_id'] = auth()->id();
+
         Idea::create($validated);
 
         return redirect()->route("dashboard")->with('success', 'Idea created succesfully');
@@ -34,12 +36,19 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() != $idea->user_id) {
+            abort(404);
+        }
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
 
     public function update(Idea $idea)
     {
+
+        if (auth()->id() != $idea->user_id) {
+            abort(404);
+        }
         $validated = request()->validate([
             'content' => 'required | min:3 | max:240',
         ]);
@@ -51,6 +60,10 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() != $idea->user_id) {
+            abort(404);
+        }
+
         $idea->delete();
 
         /*
@@ -61,6 +74,7 @@ class IdeaController extends Controller
         $idea = Idea::where('id',$id)->firstOrFail()
         $idea->delete();
         */
+
 
         return redirect()->route("dashboard")->with('success', 'Idea deleted succesfully');
     }
